==> 2 USE ME Merge in pre-fetched lyrics.ipynb <==
import pandas as pd
import numpy as np
import json

from IPython.display import Markdown as md


# 1. load in data

songs_df = pd.read_csv('./data/1 DONE RIGHT OUTPUT unique songs.csv')
songs_df

LYRIC_FOLDER_PATH = './data/lyric-fetching-datasets'

lyrics_df = pd.read_csv(f'{LYRIC_FOLDER_PATH}/{"2.6 OUTPUT lyrics for all songs"}.csv')
lyrics_df

# 2. Merge (note we'll drop ~60 tracks identified as instrumental whe creating the lyrics dataset)

### MINOR NOTE: there's seems to be duplicate with a wrong song_id. No worries, it gets dropped!

a = set(songs_df.song_id.unique())
b = set(lyrics_df.song_id.unique())
list(b.difference(a))

lyrics_df[lyrics_df.song_id.str.contains('Got You Under My Skin')].song_id.values

songs_df[songs_df.song_id.str.contains('Got You Under My Skin')].song_id.values

merged = songs_df.merge(
    right=lyrics_df,
    on='song_id',
    how='inner'
)

merged

## Sanity checks

merged.lyrics.isna().sum()

# check that Oct-Dec 22 was merged in
merged.song_id.str.contains('Lift Me UpRihanna').sum()

# 3. Export!

merged.to_csv('./data/2 OUTPUT Songs with lyrics.csv', index=False)
