==> 2.3 SUB WORK - stitch all the (re)-fetched data into one the files we can directly use to form the master dataset.ipynb <==
import pandas as pd
import json
import numpy as np

LYRIC_FOLDER_PATH = './data/lyric-fetching-datasets'

IS_INSTRUMENTAL_RETURN_FLAG = 'THIS_IS_AN_INSTRUMENTAL'

# GOAL: stitch all lyrics together, so we have a set of unique lyrics ready to be used in the main data pipeline 

ie these details should be hidden in this side notebook, since we don't need to revisit them constantly

# 1️⃣ FIRST: just load up all the data:

## ~ 58 songs identified as *instrumentals* (there may be more, potentially, which contain bogus lyrics, but haven't been detected)

instrumentals_so_no_lyrics = pd.read_csv(f'{LYRIC_FOLDER_PATH}/{"INSTRUMENTAL SONGS final"}.csv')
instrumentals_so_no_lyrics
# NOTE: I add to this in a later notebook 

## BASE DATASET: pure lyrics, but some may need to be dropped or replaced by more correct data

base_df = pd.read_csv(f'{LYRIC_FOLDER_PATH}/{"has_lyrics"}.csv')
base_df

## 69 FIXED URL LYRICS: ones I corrected by finding a bad URL, then choosing a better one. NOTE: contains 1 instrumental (that I know of)

better_url_refetched_lyrics = pd.read_csv(f'{LYRIC_FOLDER_PATH}/{"Refetched with more likely URL (good to go now!)"}.csv')
better_url_refetched_lyrics


## ~ 420 songs (~370 have lyrics, ~50 instrumental) REFETCHED FOR MISC REASONS: had characters that needed to be removed for search, maybe the call just failed randomly. NOTE: all the ~50 rows missing lyrics are instrumentals, which I've identified & saved as a seperate file

big_refetch_of_misc = pd.read_csv(
    f'{LYRIC_FOLDER_PATH}/{"DONE WOOHOO refetching about 420 non-lyric-ed songs and removing instrumentals"}.csv'
)[['song_id', 'urls', 'lyrics']]
big_refetch_of_misc

# 2️⃣ SECOND: stitch

How i propose:
0. start with the base dataset
1. UPDATE PHASE: we want to append or *replace* newly fetched data from the 2 datasets
2. FILTER OUT INSTRUMENTALS
3. check our coverage, export as lean as possible for merging

JUST_SONG_ID_AND_LYRICS_COLS = ['song_id', 'lyrics']
SONG_ID = 'song_id'

def remove_overlap(base_df, overlap_df_to_remove):
    return base_df[
        base_df.song_id.apply(
            lambda i: i not in overlap_df_to_remove.song_id.values
        )
    ]

## A. Ok, the 69 improved need to be swapped in:

better_url_plus_base_df = pd.concat([
    remove_overlap(
        base_df,
        better_url_refetched_lyrics
    ),
    better_url_refetched_lyrics[JUST_SONG_ID_AND_LYRICS_COLS]
])

better_url_plus_base_df

## B. Now add refetched ~420

big_refetch_of_misc

big_refetch_of_misc[big_refetch_of_misc.lyrics.isna()]

# to test overlap:
# better_url_plus_base_df.merge(on='song_id', right=misc_non_null) 

misc_plus_better_url_plus_base_df = pd.concat([
    remove_overlap(
        better_url_plus_base_df,
        big_refetch_of_misc
    ),
    big_refetch_of_misc
])

misc_plus_better_url_plus_base_df.drop_duplicates(subset=['song_id']).dropna(subset=['song_id', 'lyrics']) # just to prove we dont' need it :)


final_with_instrumentals_removed = remove_overlap(
    misc_plus_better_url_plus_base_df,
    instrumentals_so_no_lyrics
)
final_with_instrumentals_removed

# 3️⃣ Now just save the lyrics by `song_id`!

def save_lyrics(lyrics_subset, file_name):
    lyrics_subset.to_csv(f'{LYRIC_FOLDER_PATH}/{file_name}.csv', index=False)
    return lyrics_subset

save_lyrics(
    final_with_instrumentals_removed,
    '2.3 OUTPUT of lyrics by song id with urls'
)
