==> 2.6 SUB WORK fixes and fetching Oct - Dec 2022 songs.ipynb <==
import pandas as pd
import numpy as np

LYRIC_FOLDER_PATH = './data/lyric-fetching-datasets'
IS_INSTRUMENTAL_RETURN_FLAG = 'THIS_IS_AN_INSTRUMENTAL'

# Load up data

songs_df_from_2_output = pd.read_csv('./data/2 OUTPUT Songs with lyrics.csv') 
songs_df_from_2_output

songs_df = pd.read_csv('./data/1 DONE RIGHT OUTPUT unique songs.csv')


lyrics_df = pd.read_csv(f'{LYRIC_FOLDER_PATH}/{"2.5 OUTPUT lyrics for all songs sans instrumentals"}.csv')
lyrics_df

# Find new songs

songs_df[
    songs_df.song_id == 'Lift Me UpRihanna'
]

lyrics_df[
    lyrics_df.song_id == 'Lift Me UpRihanna'
]

merged = songs_df.merge(
    right=lyrics_df,
    on='song_id',
    how='outer', indicator=True
)

# merged[
#     merged['_merge'] == 'right_only'
# ]

new_songs = merged[
    merged.lyrics.isna() &
    merged.chart_debut.str.startswith('2022-')
]
new_songs

# Problem songs

PROBLEM_SONGS = [
    'Macarena (Bayside Boys Mix)Los Del Rio',
    'Because You Loved Me (From "Up Close & Personal")Celine Dion'
]

problem_songs = songs_df[
    songs_df.song_id.apply(lambda _id: _id in PROBLEM_SONGS)
]
problem_songs

# Fetch lyrics

import lyricsgenius
GENIUS_CLIEN_API_KEY = '8ZiCNVQqbSF6NM0nwVnVv6Y0EBR7BsVQ9HaVgPdqJOZR1l6XlymR-pUUi6z3GHPx' # be nice :)
genius = lyricsgenius.Genius(GENIUS_CLIEN_API_KEY)

BANNED_CHARACTERS = ['[', ']', '"', "'", '+', ',', '.', '(', ')', '/', '\\', '-', '&', '?', '!']
def clean_it(s):
    cleaned = s
    for naughty_char in BANNED_CHARACTERS:
        cleaned = cleaned.replace(naughty_char, '')
    return cleaned.strip()

def clean_artists(artist_str):
    s = artist_str.split('(')[0]
    s = s.split(' and ')[0]
    s = s.split(' with ')[0]
    s = s.split(' With ')[0]
    s = s.split(' And ')[0]
    s = s.split(' Featuring ')[0]
    return clean_it(s)

def get_search_term(_song, _performer):
    song = clean_it(_song.split('(')[0].split('/')[0])
    performer = clean_artists(_performer)
    return f'{performer} {song}'

def search_song_chose_first_then_get_lyrics(song, performer):
    try: 
        print(get_search_term(song, performer))
        songs = genius.search_songs(get_search_term(song, performer))
        urls = [song['result']['url'] for song in songs['hits']]
        print(urls)
        return [genius.lyrics(song_url=urls[0]), urls]
    except:
        print(f'OOPS: {performer} {song}')
        
        
def fetch_lyrics_and_urls(songs_subset_df):
    return songs_subset_df.apply(
        lambda row: search_song_chose_first_then_get_lyrics(song=row['song'], performer=row['performer']),
        axis=1
    )

songs_to_fetch = pd.concat([
    new_songs,
    problem_songs
]).reset_index(drop=True)
songs_to_fetch

res = fetch_lyrics_and_urls(songs_to_fetch)

res

# Merge new and fixed songs back into full dataset

JUST_SONG_ID_AND_LYRICS_COLS = ['song_id', 'lyrics']
SONG_ID = 'song_id'

def remove_overlap(base_df, overlap_df_to_remove):
    return base_df[
        base_df.song_id.apply(
            lambda i: i not in overlap_df_to_remove.song_id.values
        )
    ]

import json
def convert_fetched_lyrics_to_df(fetched_lyrics_and_url_options_arrays, original_df):
    just_lyrics = [x and x[0] for x in fetched_lyrics_and_url_options_arrays]
    just_stringified_urls = [x and json.dumps(x[1]) for x in fetched_lyrics_and_url_options_arrays]
    save_to_disk = pd.DataFrame({
        'song_id': original_df.song_id, 'lyrics': just_lyrics, 'urls': just_stringified_urls
    })
    return save_to_disk

fixes_and_new_fetched_df = convert_fetched_lyrics_to_df(res, songs_to_fetch)

lyrics_updated = pd.concat([
    remove_overlap(lyrics_df, fixes_and_new_fetched_df),
    fixes_and_new_fetched_df
]).reset_index(drop=True)
lyrics_updated

for a in lyrics_updated[lyrics_updated.song_id.apply(lambda i: i in PROBLEM_SONGS)].urls:
    print(a)

lyrics_updated.lyrics.isna().sum()

# Export

def save_lyrics(lyrics_subset, file_name):
    lyrics_subset.to_csv(f'./data/lyric-fetching-datasets/{file_name}.csv', index=False)
    return lyrics_subset

save_lyrics(lyrics_updated, '2.6 OUTPUT lyrics for all songs')
