==> 4. Embed lyrics VIA NOMIC.ipynb <==
import pandas as pd
import numpy as np

np.__version__

# Pull in data (prepped in other notebook)

lyric_lines_ready_for_analysis = pd.read_csv('./data/3 OUTPUT One lyric lines per row, cleaned.csv')

lyric_lines_ready_for_analysis

# 1. Create an embedding "map" with Nomic's Atlas

from nomic import AtlasClient
import numpy as np

atlas = AtlasClient()

nomic_ready_df = lyric_lines_ready_for_analysis.dropna()[['lyric_line', 'generic_genre', 'performer', 'song']].to_dict('records')

import json

nomic_ready_df = json.loads(json.dumps(nomic_ready_df))


nomic_ready_df[:100]

# Create a map:

res = atlas.map_text(
  data=nomic_ready_df,
  indexed_field='lyric_line',
  colorable_fields=['generic_genre'],
  is_public=True,
  map_name='Hot 100 Lyrics through 2022, v0.9: for text search',
  map_description='For finding story with McGhee',
  num_workers=20
)


res#['map']

### now put it back in the dataframe, with the index they snuck in (hey!)

pd.DataFrame.from_dict(nomic_ready_df)

# 2. Use Atlas as an embedding engine: pull the embedding vectors for each row (tagged by ID), merge with the data

# v0.2
project_id = '3c3fff9b-7f54-4711-bde6-7efb006fa977'
atlas_index_id = '81668f2e-32a1-4fab-a3d0-f4857c36c80c'
output_dir = './data/atlas-export'

# atlas.download_embeddings(
#     project_id, 
#     atlas_index_id, 
#     output_dir, 
#     num_workers=10
# )

## open picked files

import os
import pickle
import pandas as pd

PICKLE_REPO_PATH = './data/atlas-export/'


pkle_file = '81668f2e-32a1-4fab-a3d0-f4857c36c80c_137000_138000.pkl'
obj = pd.read_pickle(f'{PICKLE_REPO_PATH}{pkle_file}')

def get_all_pkls_as_a_single_df(path=PICKLE_REPO_PATH):
    ids_all_merged_from_pkl_files = None
    with os.scandir(path) as it:
        for entry in it:
            if entry.name.endswith(".pkl") and entry.is_file():
                if ids_all_merged_from_pkl_files is not None:
                    ids_all_merged_from_pkl_files = pd.concat([ids_all_merged_from_pkl_files, pd.DataFrame(obj)])
                else:
                    ids_all_merged_from_pkl_files = pd.DataFrame(obj)

    return ids_all_merged_from_pkl_files

get_all_pkls_as_a_single_df()

##
