==> 4. Embed lyrics.ipynb <==
import pandas as pd
import numpy as np

np.__version__

# ⚠️ BE ADVISED: this notebook is a bit slow, but thanksfully doesn't seem to crash or overheat the computer ⚠️ 

# Pull in data (prepped in other notebook)

lyric_lines_ready_for_analysis = pd.read_csv('./data/3 OUTPUT One lyric lines per row, cleaned.csv')

lyric_lines_ready_for_analysis

# EMBED WORDS USING Universal Encoder

#@title Load the Universal Sentence Encoder's TF Hub module
from absl import logging

import tensorflow as tf

import tensorflow_hub as hub
import matplotlib.pyplot as plt
import numpy as np
import os
import pandas as pd
import re
import seaborn as sns

model = hub.load('./data/universal-sentence-encoder-large_5/')

def embed(input):
    return model(input)

## Ok, so the notebook crashes trying to embed *all at once*, so let's break this up:

len(lyric_lines_ready_for_analysis)

lyrics_array = list(lyric_lines_ready_for_analysis.lyric_line.values)
# FOR TESTING:
lyrics_array = lyrics_array#[:20000]

def process_it(batch):
    return embed(batch).numpy()

def process_one_batch(batch, running_total):
    processed_batch = process_it(batch)
    return [*running_total, *processed_batch]

def batch_it(full_array, increment):
    running_total = []
    
    _max = len(full_array)
    SAFETY = 0

    lower_bound = 0
    upper_bound = lower_bound + increment

    while upper_bound < _max and SAFETY < _max:
        running_total = process_one_batch(full_array[lower_bound:upper_bound], running_total)
        lower_bound = upper_bound
        upper_bound = lower_bound + increment
        SAFETY = SAFETY + 1
        print(f'Finished: [{lower_bound} : {upper_bound}]')

    running_total = process_one_batch(full_array[lower_bound:_max], running_total)
    return running_total

res = batch_it(lyrics_array, increment=1000)
len(res)

message_embeddings = res

message_embeddings_df = pd.DataFrame(message_embeddings)
message_embeddings_df

# Export!

message_embeddings_df.to_csv('./data/4 OUTPUT 1d embeddings array.csv', index=False)
