==> 21. Calculate weight for each song based on it's popularity (rankings, durations).ipynb <==
import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)
from IPython.display import display, Markdown

import pandas as pd
import numpy as np

import json

np.__version__

USEFUL_COLUMNS = ['chart_position', 'song', 'performer', 'song_id']
hot_100_raw = pd.read_csv('./data/Hot 100 through 2022.csv')[USEFUL_COLUMNS]
hot_100_raw

SONG_RANKING_FILTER_THRESHOLD = 10
top_10 = hot_100_raw[
    hot_100_raw.chart_position <= SONG_RANKING_FILTER_THRESHOLD
].copy()
top_10['score_for_that_week'] = top_10.chart_position.apply(lambda chart_position: 11 - chart_position)
top_10

popularity_score_df = top_10.groupby(['song', 'performer']).sum()['score_for_that_week'].reset_index().rename(
    columns={'score_for_that_week': 'popularity_score'}
)
popularity_score_df.sort_values('popularity_score')

popularity_score_df.to_csv('./data/21-OUTPUT-scored-with-sum-of-popularity-over-time.csv', index=False)
