==> 22.2 v4 (part4, just prep) format and export manual love song type corrections.ipynb <==
import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)
from IPython.display import display, Markdown

import pandas as pd
import numpy as np

np.__version__

# Pull in data

PREVIOUS_OUTPUT_PATH = './data/22.2-v3-OUTPUT-data-through-sep-2023-ALL-SONGS-LABELLED-with-good-riddance-corrections.csv'
prev_df = pd.read_csv(PREVIOUS_OUTPUT_PATH)
prev_df['song_id'] = prev_df['song'] + prev_df['performer']
prev_df

# Fetch the correction gSheet
G_SHEET_URL = "https://docs.google.com/spreadsheets/d/e/2PACX-1vTwJo327jLAd-uMGiUeROIe7zrKt95XILENQuCN-K8UbAN_XCjemGXjFnzwVSfIooMmEMXzuZWLPL4v/pub?gid=0&single=true&output=csv"
gsheet_hand_corrections_df = pd.read_csv(G_SHEET_URL)

# Convert song_index to integer
gsheet_hand_corrections_df['song_index'] = gsheet_hand_corrections_df['song_index'].astype('Int64')
gsheet_hand_corrections_df

correction_df = gsheet_hand_corrections_df[['song_index', 'proposed']]
correction_df = correction_df[
    correction_df.song_index.notna()
]
CHANGED_TYPES = {
    'Pursuit': 'Courtship & Anticipation',
    'Sexual Confidence': 'Sexual Conquest',
    'Heartache': 'Longing & Heartbreak',
    'Not a Love Song': np.nan,
}
correction_df['proposed'] = correction_df['proposed'].replace(CHANGED_TYPES)
correction_df

# export

WITH_LOVE_SONG_TYPES_PATH = './data/22.2-v4-OUTPUT-just-the-hand-corrections-to-apply.csv'
correction_df.to_csv(WITH_LOVE_SONG_TYPES_PATH, index=False)
