==> 22.3 Parse performer field to split out individual performers.ipynb <==
import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)
from IPython.display import display, Markdown

import pandas as pd
import numpy as np

import json

np.__version__

# Pull in data

EXISTING__PERFORMER_SPLIT_FILE = './data/22.3-OUTPUT-gpt4-performer-to-performer-list-str-dicitonary.csv'
existing_lookup_table = pd.read_csv(EXISTING__PERFORMER_SPLIT_FILE)
existing_lookup_table

THRU_SEP_2023_PATH = './data/songs_df-through-Sep-2023.csv'
songs_df = pd.read_csv(THRU_SEP_2023_PATH)
songs_df

# make a df of all the rows from songs_df where the performer is NOT in the existing_lookup_table
df = songs_df[~songs_df['performer'].isin(existing_lookup_table['performer'])]
df

# Split `performer` field into individual performers. We want a primary performer eg `Lady Gaga` and other performers `ft Justin Bieber` but as `Justin Bieber`.

GPT_SPLIT_PROMPT = """
You are an expert pop music critic, familiar with all the musical artists of the last 70 years.
Given a string containing the performer or performers for a song, return a list of the performers in the same order they appear in the string.

Sample input:
```json
[
    "Kid Rock Featuring Sheryl Crow", 
    "Pitbull Featuring Ne-Yo, Afrojack & Nayer", 
    "The Partridge Family Starring Shirley Jones Featuring David Cassidy", 
    "FEFE ft. Nicki Minaj, Murda Beatz",
    "The Beatles With Billy Preston",
    "Johnny Mathis/Deniece Williams"
]
```

Sample output:
```json
{"output":
[
    ["Kid Rock", "Sheryl Crow"],
    ["Pitbull", "Ne-Yo", "Afrojack", "Nayer"],
    ["The Partridge Family", "Shirley Jones", "David Cassidy"],
    ["FEFE", "Nicki Minaj", "Murda Beatz"],
    ["The Beatles", "Billy Preston"],
    ["Johnny Mathis", "Deniece Williams"]
]
}
```

Ok, now do this next input:
"""

# JUST USE GPT... sigh

# Fetch new labels using GPT4

import os
from openai import OpenAI

# First, activate the './.env' file which sets the OPENAI_API_KEY environment variable using python-dotenv
from dotenv import load_dotenv
load_dotenv()

client = OpenAI(
    # This is the default and can be omitted
    api_key=os.getenv("OPENAI_API_KEY"),
)

import ast

def eval_this_str(s):
    return ast.literal_eval(s)

def get_love_song_labels_for_rows(row_subset_str):
    try: 
        completion = client.chat.completions.create(
          model="gpt-4o",
          response_format={'type': "json_object"},
          messages=[
            {"role": "user", "content": GPT_SPLIT_PROMPT},
            {"role": "user", "content": row_subset_str},
          ]
        )
        content = completion.choices[0].message.content
        # return content
        response_content = eval_this_str(content)['output']
        # print(f"Response contains {completion['usage']['completion_tokens']} tokens out of 4096 maximum allowed tokens.")
        return response_content
    except Exception as e:
        print('ERROR FETCHING', e)
        return []

### Fetch in batches

NUM_TOTAL_COLUMNS = 15 # play it safe
def fetch_all_data_in_batches(df, increment=50):
    gpt_labeled_result = []
    for i in range(101):
        subset = df['performer'][i*increment : (i+1)*increment].reset_index(drop=True)
        if len(subset):
            subset_as_json_string = subset.to_json(orient='values')
            print([i*increment, (i+1)*increment])
            print(subset_as_json_string, '\n')
            fetched_and_formatted_res = get_love_song_labels_for_rows(subset_as_json_string)
            # fetched_and_formatted_res = [['The DeFranco Family', 'Tony DeFranco']]
            # Append the original performer string as the first element in each sub array (that's effectively the id so we can merge it in later)
            fetched_and_formatted_res = [[subset[j], *fetched_and_formatted_res[j]] for j in range(len(fetched_and_formatted_res))]
            # Ok, we need all the sub arrays to have the same number of elements, so we'll pad the shorter ones with empty strings
            for j in range(len(fetched_and_formatted_res)):
                fetched_and_formatted_res[j] = fetched_and_formatted_res[j] + ['']*(NUM_TOTAL_COLUMNS - len(fetched_and_formatted_res[j]))
            print(fetched_and_formatted_res)
            gpt_labeled_result = [*gpt_labeled_result, *fetched_and_formatted_res]
    return gpt_labeled_result

def get_file_path(postfix):
    return f'./data/22.3-TEMP-gpt4-performer-split-{postfix}.csv'

def save_to_temp_csv(formatted_gpt_result, run_num):
    first_label_pass_df = pd.DataFrame(
        formatted_gpt_result, 
        # playing it safe LOL
        columns=['performer', *[f'performer_{i}' for i in range(1, NUM_TOTAL_COLUMNS)]]
    )
    first_label_pass_df.to_csv(get_file_path(run_num), index=False)

# (run a test call to be sure things look ok)
For 100 rows at once:
For GPT4 turbo: 3min, about 3400 tokens out of 4096 maximum allowed tokens 

TEST_POSTFIX = 'JUST A TEST, DELETE ME2'
save_to_temp_csv(
    fetch_all_data_in_batches(
        # unique_performers_df.head(5)
        df[df.performer.str.contains(', ')].head(5)
    ),
    TEST_POSTFIX
)

pd.read_csv(get_file_path(TEST_POSTFIX))

# Do the API calls

unique_performers = df['performer'].unique()
unique_performers_df = pd.DataFrame(unique_performers, columns=['performer'])
unique_performers_df

DATA_KEY_1 = 'through-september-2023'
# save_to_temp_csv(
#     fetch_all_data_in_batches(
#         unique_performers_df.iloc[0:1000],
#     ), 
#     DATA_KEY_1
# )

# save_to_temp_csv(
#     fetch_all_data_in_batches(
#         unique_performers_df.iloc[1000:2000],
#     ), 
#     2
# )

# save_to_temp_csv(
#     fetch_all_data_in_batches(
#         unique_performers_df.iloc[2000:3000], # there are only about 2400, to be clear
#     ), 
#     3
# )

print('DONE!')

split_by_gpt_df = pd.concat([
    pd.read_csv(get_file_path(DATA_KEY_1)),
    # pd.read_csv(get_file_path(2)),
    # pd.read_csv(get_file_path(3)),
])

# unique_performers_df get all the unique performers who are NOT in split_by_gpt_df
assert(0 == len(unique_performers_df[~unique_performers_df.performer.isin(split_by_gpt_df.performer)]))

split_by_gpt_df

# Inspect, prepare, export

# What is the maximum number of performers in a single song?
# SHow me the total count of performers in each column:
split_by_gpt_df.apply(lambda x: x.value_counts().sum())

# show me rows with non-null values in performer_5 column:
# split_by_gpt_df[split_by_gpt_df['performer_5'].notnull()]

# MANUALLY SET THIS BASED ON THE ABOVE:

MAX_PERFORMERS = 4

# create an export_ready_df where any completely empty columns are dropped
export_ready_df = split_by_gpt_df.dropna(axis=1, how='all')
export_ready_df

# get a list of all unique values in perfomer_N columns, exlcude NaN
unique_performers_after_gpt_split = pd.concat([export_ready_df[f'performer_{i}'] for i in range(1, MAX_PERFORMERS)]).value_counts()
unique_performers_after_gpt_split

# Using assert(), sanity check that the character "|" does NOT appear in any of the performer columns
assert(
    0 == sum(
        [unique_performers_after_gpt_split.index[i].count('|') for i in range(len(unique_performers_after_gpt_split))]
    )
)

# Create a new df where performer1-7 are concatinated into a "|" separated string, and performer column is left. So there are 2 columns in the new df.
string_concat_df = export_ready_df.copy()
# skip any NaN values
string_concat_df['performers_list_str'] = string_concat_df.apply(lambda x: '|'.join([x[f'performer_{i}'] for i in range(1, MAX_PERFORMERS+1) if pd.notnull(x[f'performer_{i}'])]), axis=1)
string_concat_df = string_concat_df.drop([f'performer_{i}' for i in range(1, MAX_PERFORMERS+1)], axis=1)
# append "|" to the end of each string in performers_list_str
string_concat_df['performers_list_str'] = string_concat_df['performers_list_str'].apply(lambda x: x + '|')
string_concat_df


# CONCAT FETCHED DATA WITH EXISTING DATA

concatted = pd.concat([
    existing_lookup_table,
    string_concat_df
])

# assert that the number of rows in the new df is the same as the number of rows in the original df
# assert(len(concatted) == len(df.performer.unique()))
assert(len(df.performer.unique()) + len(existing_lookup_table.performer.unique()) == len(concatted))


# Export

# save the final result
PERFORMER_SPLIT_FILE = './data/22.3-OUTPUT-gpt4-performer-to-performer-list-str-dicitonary-through-sep-2023.csv'
concatted.to_csv(PERFORMER_SPLIT_FILE, index=False)

pd.read_csv(PERFORMER_SPLIT_FILE)
