==> 22.4 Merge performers with multiple names.ipynb <==
import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)
from IPython.display import display, Markdown

import pandas as pd
import numpy as np

np.__version__

# Pull in data

PREV_PERFORMER_SPLIT_FILE = './data/22.3-OUTPUT-gpt4-performer-to-performer-list-str-dicitonary-through-sep-2023.csv'
unmerged_df = pd.read_csv(PREV_PERFORMER_SPLIT_FILE)
unmerged_df

# General approach:

It won't necessarily be perfect. 

However, high accuracy pretty much only matters for *very popular* performers.
1. We'll likely run into them as we test the project (we have)
2. The main purpose is to be more accurate when comparing, say, Beyonce vs GaGa. If there is a Joe Shmmo and a Joseph Schmo, but he is a random with 2 songs, not merging him will likely not even be relevent to the big artists we'll reference agg stats for.

### 1. Apply manually identified merges 
This could probably be a gSheet that anyone can add to, and I can just read from here.

### 2. Low hanging fruit: finding mismatching case & small Levenshtien distances

### 3. OPTIONAL: now that we've done most of what we can, try embedding & visualizing in 2D (maybe Nomic makes this easy?)

We could scan visually in the viz.

But we could also programmatically pull up in a table any pairs who are close to each other.

# 1. Manually identified merges

# Pull from gSheet
G_SHEET_URL = "https://docs.google.com/spreadsheets/d/1RUeuDF6yhv8xnoPjQoNa9hXsYDfNQzuOaWkSyww6Vi8/export?format=csv&gid=0"
g_sheet_manual_merges_df = pd.read_csv(G_SHEET_URL)
g_sheet_manual_merges_df

# Create a copy of unmerged_df
merged_df = unmerged_df.copy()

# DROP MISC
# Replace any sub strings reading "his Orchestra|" or "His Orchestra|" with ""
merged_df['performers_list_str'] = merged_df['performers_list_str'].str.replace("|his Orchestra|", "|", regex=False)
merged_df['performers_list_str'] = merged_df['performers_list_str'].str.replace("|His Orchestra|", "|", regex=False)

# Iterate over each row in g_sheet_manual_merges_df
for index, row in g_sheet_manual_merges_df.iterrows():
    for col in ['name_1', 'name_2', 'name_3', 'name_4']:
        name_to_replace = row[col]
        if pd.notna(name_to_replace):
            # Find names to replace by searching for them as substrings of performers_list_str
            merged_df['performers_list_str'] = merged_df['performers_list_str'].str.replace(f"{name_to_replace}|", f"{row['final_name_to_use']}|", regex=False)


# Check that it worked:
for index, row in g_sheet_manual_merges_df.iterrows():
    for col in ['name_1', 'name_2', 'name_3', 'name_4']:
        name_to_replace = row[col]
        if pd.notna(name_to_replace) and name_to_replace != row['final_name_to_use']:
            assert(merged_df['performers_list_str'].str.contains(f"{name_to_replace}|", regex=False).sum() == 0)



# 2. Verify the Low hanging fruit: mismatched case and Levenshtein distance

## Verify that there are no overlooked merges based purely on case, eg "Nas" and "NAS"

# Parse using "|" the performers_list_str back into a list, and combine them into one big, array of unique values
performers_list = merged_df['performers_list_str'].str.split("|").apply(lambda x: [i for i in x if i]).values
performers_list = np.concatenate(performers_list)
performers_list = np.unique(performers_list)

# (to inspect any found duplicates)
merged_df[
    merged_df.performers_list_str.str.contains("randy vanwarmer|", regex=False, case=False)
]

# Now lower case all the names, and see if the length changes
performers_list_lower = np.char.lower(performers_list)

# VERIFY: Now, print out any names that appear more than once
duplicates = pd.Series(performers_list_lower).value_counts()
duplicates = duplicates[duplicates > 1]
duplicates
assert(len(duplicates) == 0)

## Check for almost identical names

from Levenshtein import distance

# Now, compare the distance(performer1, performer2) between every single performer
# Filter to only those that are within a certain distance
MIN_DISTANCE = 2
performer_distances = []
for i, performer1 in enumerate(performers_list):
    for j, performer2 in enumerate(performers_list):
        if i < j:
            if distance(performer1, performer2) <= MIN_DISTANCE:
                performer_distances.append((performer1, performer2))

# Now, print out the performer_distances
performer_distances


# Now filter performer_distances to just names that are more than X letters
performer_distances_filtered = []
MIN_LENGTH = 5
for performer1, performer2 in performer_distances:
    if len(performer1) > MIN_LENGTH and len(performer2) > MIN_LENGTH:
        performer_distances_filtered.append((performer1, performer2))

performer_distances_filtered

# Ok, looks pretty good, export time!

PERFORMER_SPLIT_FILE = "./data/22.4-OUTPUT-performer_list_str-with-multi-name-artists-merged.csv"
merged_df.to_csv('./data/22.4-OUTPUT-performer_list_str-with-multi-name-artists-merged.csv', index=False)
